Module Sc

module Init : sig ... end

Interface for initializing the tools by calling the functors with the correct configuration. Each tool listed in this module (see `to_init`) will be set up with the configuration provided by the command line or the configuration file.

module Main : sig ... end

module Version : sig ... end