Module Sc.Init

Interface for initializing the tools by calling the functors with the correct configuration. Each tool listed in this module (see `to_init`) will be set up with the configuration provided by the command line or the configuration file.

module type GenericInit = sig ... end

The signature of a tool to initialize.

module type SimpleInit = GenericInit with type label_family := [ `simple ]

module type AnyInit = GenericInit with type label_family := Sc_lib.Types.Label.any

module type InitWithOpts = sig ... end

type init_data = {

config : (module Sc_lib.Config.T);
(*
The configuration of the project.
*)

staticdb_params : (module Sc_lib.Tool.Opt with type t = Sc_staticdb.Main.TomlSpec.opt);
(*
The parameters of the static database, which is today handled by the lannot file.
*)

tools : (module InitWithOpts) list;
(*
The tools to initialize.
*)

close_logs : unit -> unit;
(*
Closes the log channel, opened by the function generating this init_data value.
*)

}

All the modules needed to interact with the different tools after initialization.

val setup : ?enable_console_timing:bool -> ?enable_logfile:bool -> ?only_with_config_file:[ `toml ] Sc_sys.File.TYPES.file -> unit -> init_data Lwt.t

Initializes the different modules after reading the command line arguments (argv) and the content of the config file ("config.toml"). Param `enable_console_timing` makes the output cleaner for the console. Param `enable_logfile` redirects the logs to a file "x.log" where `x` is run number. Param `only_with_config_file` specifies a file to use as the only source of configuration for the run.

val init_tools : (module Sc_lib.Ctypes_bridge.Struct.REPR with type Val.t = 'r) -> init_data -> ((Sc_lib.Types.project_data, Sc_shareddb.Main.working_data, 'r Sc_shareddb.Main.runtime_data) Sc_lib.Tool.tool_accessor, string) Stdlib.result Lwt.t

Initializes the tools with the unified C types representation and configuration. Returns a tool accessor (see Sc_lib.Tool for a complete documentation of tool accessors).

val pp_config_docs : Stdlib.Format.formatter -> ('a, 'b, 'c) Sc_lib.Tool.tool_accessor -> unit

Prints the configuration options of the tool accessor.