Module TestRepr.Val

include Sc_lib.Ctypes_bridge.VAL with type typ := Sc_lib.Ctypes_bridge.Struct.typ and type value := Sc_lib.Ctypes_bridge.Struct.value and module Encoding.Params = Params

module Encoding : Sc_lib.Ctypes_bridge.ENCODING with module Params = Params

module Printer : Sc_lib.Ctypes_bridge.VALUE_PRINTER

type t

val blank : Sc_lib.Ctypes_bridge.Struct.typ -> t

blank typ returns a new value of type typ with all bytes set to null (zero).

val from_string : Sc_lib.Ctypes_bridge.Struct.typ -> string -> t

from_string typ str interprets the string of characters str as a C structured value of type typ. Raises InvalidValueRepresentation if str is not long-enough to hold such a value.

val to_string : t -> string

to_string typ v returns the C structured value v, of type typ, as a string of bytes encoded in a string.

val read : Sc_lib.Ctypes_bridge.Struct.typ -> Stdlib.in_channel -> t

read typ ic acts like from_string, but reads bytes from the input channel ic.

val write : Stdlib.out_channel -> t -> unit

write oc v writes the binary representation of the C structured value v onto oc.

val typ : t -> Sc_lib.Ctypes_bridge.Struct.typ

typ v returns the type of v.

val value : t -> Sc_lib.Ctypes_bridge.Struct.value

value v returns the value of v.

val print : Stdlib.Format.formatter -> t -> unit

print fmt v prints the structured value v as a C-like literal using the formatter fmt; in the presence of pointers, this does not generate parseable C code.

val as_c_literal : ?lit_heap:Printer.LiteralHeap.t as 'm -> t -> string * 'm

as_c_literal ?lit_heap v constructs a parseable C literal representation for the value v, along with a possibly empty set of object literals representing dynamci memory cells.

val assign_from_literal : Sc_lib.Ctypes_bridge.typdecls -> t -> Sc_lib.Ctypes_bridge.literal_binding -> unit

assign_from_literal v bindings performs a deep assignent of v w.r.t the bindings described using bindings.

val fields_as_c_literals : t -> Printer.LiteralMemory.t

fields_as_c_literals v returns a pair (fields, heap): fields is a list of pairs of strings (fname, fval), where fname is the name of a field in the type of v, and fval is a human-readable, C-parseable, representation of its value in v; in turn, heap is a possibly empty literal heap.

val write_bout : Stdlib.out_channel -> t -> unit

write_bout oc v writes on oc the fields of v as a series of binary objects in klee's KTEST file format (a.k.a BOUT).

val read_bout : Stdlib.in_channel -> Sc_lib.Ctypes_bridge.Struct.typ -> t

read_bout ic typ reads from ic a series of named binary objects in klee's KTEST file forma (a.k.a BOUT), and returns a C structure value of type typ.

May raise End_of_file in case the data is incomplete.

Raises InvalidValueRepresentation when the input data does not correspond to a value of type typ.