Module Init.Make

Parameters

module Conf : Sc_lib.Config.T

module Database : Sc_lib.Tool.Database

Signature

include Sc_lib.Tool.ToolInit with type input_data = Database.Runtime.working_data with type opt = TomlSpec.opt

module Log : Logs.LOG

include Sc_lib.Tool.TomlSpec with type opt = TomlSpec.opt

type opt = TomlSpec.opt

Options for the tool

val tool_name : string

The name of the tool

val toml_schema : opt Sc_lib.Eztoml.schema

Tool options, retrieved from project configuration

val default_opt : opt

A default option value

type input_data = Database.Runtime.working_data

type working_data

val tool_project_dir : Sc_sys.File.TYPES.dir

The directory where the tool outputs its files

val is_installed : unit -> bool Lwt.t

Checks that the tool is properly installed

val preprocess_file : opt:opt -> input_data -> working_data Lwt.t

Preprocess the file given as argument, and return operational data for the tool.

val no_preprocess_file : opt:opt -> input_data -> working_data Lwt.t

Same as `preprocess_file`, but does not preprocess file.

type label_family = label_family

The label family treated by the tool

type runtime_data = Database.Runtime.runtime_data

The shared data

val tool_kind : [ `Preprocess | `Static | `Dynamic ]

The tool kind.

val generate_test_cases : opt:opt -> shareddb:runtime_data -> working_data -> label_family Sc_lib.Types.Label.t list -> unit Sc_sys.Process.TYPES.process Lwt.t

Starts the tool for generating test cases. For each proof objective, returns a list of proof_objective id and the status the tool concluded on.

TODO: do not return test cases anymore, only proof objectives status.