Module Printer.LiteralHeap

Literal representation of heaps

type t

The type of literal heaps, which describes a set of non-overlapping memory cells that can each be individually addressed.

val empty : t

The empty literal heap

val print : ?static:bool -> t Basics.PPrt.pp

print ?static fmt h uses fmt to print parseable C code that declares and initializes all heap memory cells in h. Initialized declarations (i.e, `typ cell = lit;') are used if static = true; calls to `memcpy` are used otherwise (i.e, `typ cell; (void) memcpy (&cell, ...);`).