Module Sc.Types

type logs_config = {

log_level : Logs.level option;
(*
Log level.
*)

}

type options = {

clean_start : bool;
(*
When set, start by archiving the work done on this project.
*)

config_file : [ `toml ] Sc_sys.File.TYPES.file option;
(*
Configuration file given on the command-line.
*)

amending_table : Toml.Types.table;
(*
Amendments to the configuration that come from command-line arguments.
*)

print_statistics : bool;
(*
When set, print statistics at the end of the run.
*)

}

Command-line-only options

type check_options = {

check_initialization : bool;
(*
Additionally perform project initialization.
*)

}

type dump_doc_options = {

format : [ `Rst | `Formatted ];

destination : [ `Stdout | `Filename of string ];

}